too much stuff; mostly attempts to refactor; i am trying to remove circular dependencies; there will be more
